dock: add elantec_touchscreen.idc from tf700 flavour
A CM11 install on the tf701t will almost work with the touchpad on the
keyboard dock except that, annoyingly, the touchpad is assumed to be an
actual touch screen device and so one ends up with the "circle"-style
mouse pointer.

Copying this file from the tf700 CM configuration results in the
expected "arrow"-style mouse pointer.

-added in to PRODUCT_COPY_FILES.

Change-Id: I1383303f9aff4e2ded73b4419e0734847fd2ebd5